Start HexagonPatch Implementation
I want to only kind-of port the HexagonPatch implementation I made in
Javascript. Only kind-of because the approach always remained a bit
hacky and contained some topological bugs, too.

This here starts my adventures with Godot and C# and so far, with a lot
of "C# <java-thing>" googling, it's going quite well.

So we keep the general approach of using "hexagon patches" which are lot
like Oskar Stålberg did in "TownScaper" but not smaller Hexagons in big
Hexagons but smaller hexagons in a hexagon grid of fixed size.

We start with a 10 by 10 grid of hexagons made out of triangle fans. We
make sure that our algorithm produces the *exact* same integer
coordinates for all the shared vertices. Turns out the easiest way to do
that is to limit the radius of the hexagons to sizes for which

 Width = Round(Sqrt(3) * size)

produces an even number. A bit funky but it works.

Supported by this precision this code implements the first stage of
triangle creation and half edge structure weaving. It works on the faces
produced for the grid previously, twinning each new face with the
appropriate faces to the left and top.

The code contains a Debug.Assert that checks the actual twinned half
edges for correctness of position coordinates.

Adds a Nunit test that checks the produces list of faces for
correctness. Since we already check with the previous assert that the
coordinates of each pair of twin is right (and their Vector3 position
instances are also identical), we just need to make sure that each
triange fan blade of each Hexagon triangle fan is twinned or not
depending on whether it touches the patch border or not.